Refactor resources module to have an explicit utils file
