+++ pennylane/resource/specs.py
@@ -23,43 +23,23 @@ import time
 import warnings
 from collections import defaultdict
 from collections.abc import Callable, Iterable
-from functools import partial, wraps
+from functools import partial
 from pathlib import Path
-from typing import TYPE_CHECKING
 
 import pennylane as qp
 
-from .mlir_specs import make_level_name_unique, resources_from_analysis_pass
+from .mlir_specs import resources_from_analysis_pass
 from .resource import CircuitSpecs, SpecsResources, resources_from_tape
-
-if TYPE_CHECKING:
-    from pennylane.transforms.core import CompilePipeline
+from .utils import (
+    apply_partial_args,
+    get_last_tape_transform_level,
+    make_level_name_unique,
+    preprocess_level_input,
+    unwrap_partial,
+)
 
 # Used for device-level qjit resource tracking
 _RESOURCE_TRACKING_PREFIX = "pennylane_specs_qjit_resources"
-
-
-def _unwrap_partial(fn):
-    """Return the base callable and arguments bound by nested ``functools.partial`` wrappers."""
-    args = ()
-    kwargs = {}
-    while isinstance(fn, partial):
-        args = fn.args + args
-        kwargs = {**(fn.keywords or {}), **kwargs}
-        fn = fn.func
-    return fn, args, kwargs
-
-
-def _apply_partial_args(fn, args, kwargs):
-    """Return a callable that prepends partial-bound arguments to call-time arguments."""
-    if not args and not kwargs:
-        return fn
-
-    @wraps(fn)
-    def wrapper(*call_args, **call_kwargs):
-        return fn(*args, *call_args, **{**kwargs, **call_kwargs})
-
-    return wrapper
 
 
 def _specs_qnode(qnode, level, compute_depth, *args, **kwargs) -> CircuitSpecs:
@@ -137,90 +117,6 @@ def _specs_qjit_device_level_tracking(
         )
 
 
-def _get_last_tape_transform_level(compile_pipeline: CompilePipeline) -> int:
-    """Helper function to get the last level which is a tape transform and not an MLIR pass.
-
-    Note that this includes an implicit level 0 which corresponds to the original circuit.
-
-    Args:
-        compile_pipeline: The compile pipeline of the QNode, which contains both user-applied tape transforms and MLIR passes
-
-    Returns:
-        int: The last level which is a tape transform and not an MLIR pass, or 0 if there are no tape transforms
-    """
-    # Find the seam where transforms end and MLIR passes begin
-    # If the pass name is None, it indicates a transform which is NOT also a Catalyst pass
-    for i, trans in reversed(list(enumerate(compile_pipeline))):
-        if trans.pass_name is None:
-            #  Add 1 to account for the implicit "Before Tape Transforms" at level=0
-            return i + 1
-    return 0
-
-
-def _preprocess_level_input(  # pylint: disable=too-many-branches
-    level: str | int | slice | list[int | str],
-    marker_to_level: dict[str, int],
-    pipeline_len: int,
-    num_tape_levels: int,
-) -> list[int]:
-    """Preprocesses the level input to always return a sorted list of integers.
-
-    Args:
-        level (str | int | slice | iter[int | str]): The level input to preprocess
-        marker_to_level (dict[str, int]): Mapping from marker names to their associated level numbers.
-            Note that this should already account for any inserted lowering pass.
-        pipeline_len (int): The length of the compile pipeline (number of transforms and passes)
-        num_tape_levels (int): The number of tape levels in the compile pipeline (including the implicit level 0)
-    Returns:
-        list[int]: The preprocessed level input
-    """
-    # Account for "Before MLIR passes" level
-    total_levels = pipeline_len + 1
-
-    if num_tape_levels > 1:
-        # Account for an additional "Before Tape Transforms" level
-        total_levels += 1
-
-    if level == "all":
-        return list(range(0, total_levels))
-    if level == "all-mlir":
-        return list(range(num_tape_levels, total_levels))
-    if level == "user":
-        return [total_levels - 1]
-
-    if isinstance(level, (int, str)):
-        level = [level]
-    elif isinstance(level, slice):
-        level = list(range(level.start or 0, level.stop, level.step or 1))
-    else:
-        level = list(level)
-
-    # Convert marker names to the associated level number
-    for i, lvl in enumerate(level):
-        if isinstance(lvl, str):
-            if lvl not in marker_to_level:
-                raise ValueError(f"Marker name '{lvl}' not found in the compile pipeline.")
-            level[i] = marker_to_level[lvl]
-        elif isinstance(lvl, int):
-            if lvl < 0 or lvl >= total_levels:
-                raise ValueError(
-                    "The 'level' argument to qp.specs for QJIT'd QNodes is out of bounds, "
-                    f"got {lvl}."
-                )
-        else:
-            raise ValueError(f"Invalid level '{lvl}' in level list, expected int or str.")
-
-    level_sorted = sorted(set(level))
-    if level != level_sorted:
-        warnings.warn(
-            "The 'level' argument to qp.specs for QJIT'd QNodes has been sorted to be in ascending "
-            "order with no duplicate levels.",
-            UserWarning,
-        )
-
-    return level_sorted
-
-
 def _specs_qjit_intermediate_passes(qjit, original_qnode, level, *args, **kwargs) -> tuple[
     SpecsResources | list[SpecsResources] | dict[str, SpecsResources | list[SpecsResources]],
     str | dict[int, str],
@@ -231,7 +127,7 @@ def _specs_qjit_intermediate_passes(qjit, original_qnode, level, *args, **kwargs
     compile_pipeline = original_qnode.compile_pipeline
 
     # This value is used to determine the last level which is a transform and not an MLIR pass
-    num_tape_levels = _get_last_tape_transform_level(compile_pipeline)
+    num_tape_levels = get_last_tape_transform_level(compile_pipeline)
     if num_tape_levels != 0:
         # Account for the "Before Tape Transforms" tape at level 0
         num_tape_levels += 1
@@ -256,7 +152,7 @@ def _specs_qjit_intermediate_passes(qjit, original_qnode, level, *args, **kwargs
         "all",
         "all-mlir",
     )
-    level = _preprocess_level_input(level, marker_to_level, len(compile_pipeline), num_tape_levels)
+    level = preprocess_level_input(level, marker_to_level, len(compile_pipeline), num_tape_levels)
     level_to_name: dict[int, str] = {}  # This will be a map of level to its name
 
     tape_levels = [lvl for lvl in level if lvl < num_tape_levels]
@@ -855,7 +751,7 @@ def specs(
     # pylint: disable=import-outside-toplevel
     # Have to import locally to prevent circular imports as well as accounting for Catalyst not being installed
 
-    qnode, partial_args, partial_kwargs = _unwrap_partial(qnode)
+    qnode, partial_args, partial_kwargs = unwrap_partial(qnode)
 
     specs_fn = _specs_qnode if isinstance(qnode, qp.QNode) else None
 
@@ -879,7 +775,7 @@ def specs(
             pass
 
     if specs_fn is not None:
-        return _apply_partial_args(
+        return apply_partial_args(
             partial(specs_fn, qnode, level, compute_depth), partial_args, partial_kwargs
         )
