+++ tests/resource/test_specs.py
@@ -22,10 +22,6 @@ import pennylane as qp
 from pennylane import numpy as pnp
 from pennylane.core.shots import Shots
 from pennylane.resource import SpecsResources
-from pennylane.resource.specs import (
-    _get_last_tape_transform_level,
-    _preprocess_level_input,
-)
 
 devices_list = [
     (qp.device("default.qubit"), None),
@@ -44,98 +40,6 @@ def test_error_with_bad_key(key):
     out = qp.specs(c)()
     with pytest.raises(KeyError):
         _ = out[key]
-
-
-@pytest.mark.parametrize(
-    "level,output,expect_warnings",
-    [
-        (0, [0], False),
-        (slice(3), [0, 1, 2], False),
-        (slice(1, 3), [1, 2], False),
-        (slice(1, 4, 2), [1, 3], False),
-        ([0, 1], [0, 1], False),
-        ([0, 1, 1, 1], [0, 1], True),
-        ((0, 1), [0, 1], False),
-        (range(3, 0, -1), [1, 2, 3], True),
-        ("foo", [2], False),
-        (["foo", "bar"], [2, 3], False),
-        ((1, "foo", "baz", 4, "bar"), [1, 2, 3, 4, 5], True),
-        ("all", [0, 1, 2, 3, 4, 5, 6], False),
-        ("all-mlir", [4, 5, 6], False),
-        ("user", [6], False),
-    ],
-)
-def test_preprocess_levels(level, output, expect_warnings):
-    """Test that _preprocess_level_input works correctly"""
-    marker_to_level = {
-        "foo": 2,
-        "bar": 3,
-        # Treat MLIR lowering as level 4
-        "baz": 5,
-    }
-
-    if expect_warnings:
-        with pytest.warns(
-            UserWarning,
-            match="The 'level' argument to qp.specs for QJIT'd QNodes has been sorted to be in ascending "
-            "order with no duplicate levels.",
-        ):
-            assert _preprocess_level_input(level, marker_to_level, 5, 4) == output
-    else:
-        assert _preprocess_level_input(level, marker_to_level, 5, 4) == output
-
-
-@pytest.mark.parametrize(
-    "num_tapes, expected",
-    [
-        (  # If there are no tape transforms, the "Before Tape Transforms" level should be skipped
-            0,
-            list(range(5)),
-        ),
-        (2, list(range(6))),
-        (5, list(range(6))),
-    ],
-)
-def test_preprocess_levels_all(num_tapes, expected):
-    # Assume there are always 4 transforms in the pipeline
-    assert _preprocess_level_input("all", {}, 4, num_tapes) == expected
-
-
-def test_preprocess_levels_invalid():
-    with pytest.raises(ValueError, match="out of bounds"):
-        _preprocess_level_input(-10, {}, 5, 0)
-
-    with pytest.raises(ValueError, match="out of bounds"):
-        _preprocess_level_input(10, {}, 5, 0)
-
-    with pytest.raises(ValueError, match="Invalid level"):
-        _preprocess_level_input([1, 2, 3.14], {}, 5, 0)
-
-    with pytest.raises(ValueError, match="Marker name 'foo' not found"):
-        _preprocess_level_input("foo", {}, 5, 0)
-
-
-def test_get_last_tape_transform_level():
-    """Test that _get_last_tape_transform_level works correctly"""
-
-    @qp.transform
-    def dummy_transform(tape):
-        return (tape,), lambda res: res[0]
-
-    # If there are no transforms, the last transform level should be 0
-    assert _get_last_tape_transform_level(qp.CompilePipeline()) == 0
-    # If there are *any* tape transforms, this should return the number of tape transforms
-    # since there is an implied level 0 for "Before Tape Transforms"
-    assert _get_last_tape_transform_level(qp.CompilePipeline(dummy_transform)) == 1
-    assert _get_last_tape_transform_level(qp.CompilePipeline(dummy_transform, dummy_transform)) == 2
-
-    # MLIR passes should not be counted
-    assert (
-        _get_last_tape_transform_level(
-            qp.CompilePipeline(dummy_transform, qp.transform(pass_name="cancel_inverses"))
-        )
-        == 1
-    )
 
 
 @pytest.mark.usefixtures("enable_and_disable_graph_decomp")
